+++ oxrdflib/_types.py
@@ -1,0 +1,8 @@
+from typing import Tuple, Optional
+
+from rdflib import Graph
+from rdflib.term import Node
+
+_Triple = Tuple[Node, Node, Node]
+_Quad = Tuple[Node, Node, Node, Graph]
+_TriplePattern = Tuple[Optional[Node], Optional[Node], Optional[Node]]
